# contracts/rust/openapi.v2.yaml
openapi: 3.1.0
info:
  title: AIWF Accel Rust API
  version: 0.3.0
servers:
  - url: http://127.0.0.1:18082
paths:
...
  /tasks/{task_id}:
    get:
      summary: Query async task
      operationId: getTask
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: task state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskState"
        "404":
          description: task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskNotFoundResp"
  /tasks/{task_id}/cancel:
    post:
      summary: Cancel async task if queued/running
      operationId: cancelTask
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: cancel result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskCancelResp"
        "404":
          description: task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskNotFoundResp"
components:
  schemas:
...
    TaskState:
      type: object
      additionalProperties: false
      required: [task_id, tenant_id, operator, status, created_at, updated_at]
      properties:
        task_id:
          type: string
        tenant_id:
          type: string
        operator:
          type: string
        status:
          type: string
          enum: [queued, running, done, failed, cancelled]
        created_at:
          type: string
        updated_at:
          type: string
        result:
          type: object
          additionalProperties: true
        error:
          type: string
    TaskCancelResp:
      type: object
      additionalProperties: false
      required: [ok, task_id, status, cancelled]
      properties:
        ok:
          type: boolean
        task_id:
          type: string
        status:
          type: string
        cancelled:
          type: boolean
    TaskNotFoundResp:
      type: object
      additionalProperties: false
      required: [ok, error, task_id]
      properties:
        ok:
          type: boolean
        error:
          type: string
        task_id:
          type: string
...